Revert "wip"
This reverts commit 79ccf800ee106201dc1f76d9538227875af4bd32.

+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/storage/blob/APISpec.groovy
@@ -169,7 +169,9 @@ class APISpec extends Specification {
     def setup() {
         String fullTestName = getFullTestName(specificationContext.getCurrentIteration())
         String className = specificationContext.getCurrentSpec().getName()
-        this.testName = fullTestName
+        int iterationIndex = fullTestName.lastIndexOf("[")
+        int substringIndex = (int) Math.min((iterationIndex != -1) ? iterationIndex : fullTestName.length(), 50)
+        this.testName = fullTestName.substring(0, substringIndex)
         this.interceptorManager = new InterceptorManager(className + fullTestName, testMode)
         this.resourceNamer = new TestResourceNamer(className + testName, testMode, interceptorManager.getRecordedData())
 
@@ -200,26 +202,28 @@ class APISpec extends Specification {
             return fullName
         }
         def prefix = fullName
-        def suffix = iterationInfo.getIterationIndex()
+        def suffix = "[" + iterationInfo.getIterationIndex() + "]"
 
         return prefix + suffix
     }
 
     def cleanup() {
-        def cleanupClient = getServiceClientBuilder(primaryCredential,
-            String.format(defaultEndpointTemplate, primaryCredential.getAccountName()), null)
-            .retryOptions(new RequestRetryOptions(RetryPolicyType.FIXED, 3, 60, 1000, 1000, null))
-            .buildClient()
+        if (!liveMode()) {
+            def cleanupClient = getServiceClientBuilder(primaryCredential,
+                String.format(defaultEndpointTemplate, primaryCredential.getAccountName()), null)
+                .retryOptions(new RequestRetryOptions(RetryPolicyType.FIXED, 3, 60, 1000, 1000, null))
+                .buildClient()
 
-        def options = new ListBlobContainersOptions().setPrefix(containerPrefix + testName)
-        for (def container : cleanupClient.listBlobContainers(options, null)) {
-            def containerClient = primaryBlobServiceClient.getBlobContainerClient(container.getName())
+            def options = new ListBlobContainersOptions().setPrefix(containerPrefix + testName)
+            for (def container : cleanupClient.listBlobContainers(options, null)) {
+                def containerClient = primaryBlobServiceClient.getBlobContainerClient(container.getName())
 
-            if (container.getProperties().getLeaseState() == LeaseStateType.LEASED) {
-                createLeaseClient(containerClient).breakLeaseWithResponse(new BlobBreakLeaseOptions().setBreakPeriod(Duration.ofSeconds(0)), null, null)
+                if (container.getProperties().getLeaseState() == LeaseStateType.LEASED) {
+                    createLeaseClient(containerClient).breakLeaseWithResponse(new BlobBreakLeaseOptions().setBreakPeriod(Duration.ofSeconds(0)), null, null)
+                }
+
+                containerClient.delete()
             }
-
-            containerClient.delete()
         }
 
         interceptorManager.close()
